Simplified libfive::Tree

This is a fun one!

For a long time, math expressions were built from `libfive::Tree` objects, which
were a flyweight wrapper around a common expression cache.  This had the
advantage that expressions would only be created once:
```
auto a = libfive::X() + 1;
auto b = libfive::X() + 1;
// a == b
```

The downside is that this requires a global shared cache,
which in turn requires some
[funky stuff](https://github.com/libfive/libfive/blob/7853ff0ed6c55d099f5e54fac6bc148fcbcb82ce/libfive/src/tree/cache.cpp#L16-L31).

In addition, a fair number of the math operations are
[accidentally quadratic](https://accidentallyquadratic.tumblr.com/).
Commutative operations are badly behaved because they rebalance the math tree at
every operation:
```
(define x (lambda-shape (x y z) x))
(let f ((x x) (i 10))
  (if (= i 0) x
    (* (+ x i) (f x (- i 1)))))
```
(increase `i` from 10 to 1000 and see what happens)

A second example of quadratic behavior is remapping, which rebuilds the entire
tree:
```
(let f ((g (lambda-shape (x y z) x)) (i 10))
  (if (= i 0) g
    (f (remap-shape (g x y z) (cos x) y z) (- i 1))))
```

These operations are also stack-recursive, so sufficiently deep trees can cause
them to explode.

--------------------------------------------------------------------------------

This PR rewrites how the `libfive::Tree` object works.

There are three high-level goals:
- No global cache
- Every normal `Tree` operation should be `O(1)`
- No recursion on the stack

Trees are **no longer implicitly deduplicated**:
```
auto a = libfive::X() + 1;
auto b = libfive::X() + 1;
// a != b
```
This lets us remove the global `libfive::Cache` state, and its weird `shutdown`
functions.

In addition, tree optimizations during normal operations are limited to local
identities (i.e. `X + 0 ==> X`), without stuff like balancing of commutative
functions.

Under the hood, the tree is now using `std::variant` to represent the different
types (constant / unary / binary / oracle / etc), and I've bumped the minimum
C++ version up to C++17 (welcome to the future!).

--------------------------------------------------------------------------------

Of course, we still want deduplicated trees for evaluation.  This is handled
with `Tree::optimized()`, which deduplicates and collapses affine operations
within a `Tree`.  There's a flag in the `Tree` class which records if it was
produced by `optimized` in the first place, to make things more efficient.

Performance-sensitive code which constructs multiple evaluators will want to
store output of `Tree::optimized()`, rather than doing the optimization each
time an evaluator is constructed.

--------------------------------------------------------------------------------

Finally, this PR also make all `Tree` operations recurse on the heap, rather
than the stack.  This leads to some
[awkward code](https://github.com/libfive/libfive/blob/1cfab578c396ce5f1c52d7067011380f3200eabd/libfive/src/tree/tree.cpp#L667-L799),
but means that it's nigh-impossible to cause stack overflows just by making
really deep trees (e.g. #263)

+++ libfive/src/eval/eval_array.cpp
@@ -21,8 +21,8 @@ ArrayEvaluator::ArrayEvaluator(const Tree& root)
     // Nothing to do here
 }
 
-ArrayEvaluator::ArrayEvaluator(
-        const Tree& root, const std::map<Tree::Id, float>& vars)
+ArrayEvaluator::ArrayEvaluator(const Tree& root,
+                               const std::map<Tree::Id, float>& vars)
     : ArrayEvaluator(std::make_shared<Deck>(root), vars)
 {
     // Nothing to do here
@@ -245,7 +245,7 @@ void ArrayEvaluator::operator()(Opcode::Opcode op, Clause::Id id,
         case Opcode::OP_ATAN2:
             for (auto i=0; i < a.size(); ++i)
             {
-                out(i) = atan2(a(i), b(i));
+                out(i) = atan2f(a(i), b(i));
             }
             break;
         case Opcode::OP_POW:
@@ -261,7 +261,7 @@ void ArrayEvaluator::operator()(Opcode::Opcode op, Clause::Id id,
                             Interval(a(i), a(i)),
                             Interval(b(i), b(i))).lower();
                 else
-                    out(i) = pow(a(i), 1.0f/b(i));
+                    out(i) = powf(a(i), 1.0f/b(i));
             }
             break;
         case Opcode::OP_MOD:
