+++ libfive/src/eval/eval_deriv_array.cpp
@@ -19,8 +19,8 @@ DerivArrayEvaluator::DerivArrayEvaluator(const Tree& root)
     // Nothing to do here
 }
 
-DerivArrayEvaluator::DerivArrayEvaluator(
-        const Tree& root, const std::map<Tree::Id, float>& vars)
+DerivArrayEvaluator::DerivArrayEvaluator(const Tree& root,
+                                         const std::map<Tree::Id, float>& vars)
     : DerivArrayEvaluator(std::make_shared<Deck>(root), vars)
 {
     // Nothing to do here
@@ -166,7 +166,7 @@ void DerivArrayEvaluator::operator()(Opcode::Opcode op, Clause::Id id,
         case Opcode::OP_NTH_ROOT:
             for (Eigen::Index i=0; i < od.cols(); ++i)
                 od.col(i) = (ad.col(i) == 0)
-                    .select(0, ad.col(i) * (pow(av(i), 1.0f / bv(i) - 1) / bv(i)));
+                    .select(0, ad.col(i) * (powf(av(i), 1.0f / bv(i) - 1) / bv(i)));
             break;
         case Opcode::OP_MOD:
             od = ad;
